# gestion_canditature/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "endPoind ",
        "version": "1.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/formation/candidat/{formationId}": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Créer une candidature",
                "description": "Crée une candidature pour une formation donnée.",
                "operationId": "21b87c199871eb1c6ea68f996ec17fb7",
                "parameters": [
                    {
                        "name": "formationId",
                        "in": "path",
                        "description": "ID de la formation pour laquelle la candidature est créée",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidature créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Votre Candidature est prise en compte"
                                        },
                                        "user": {
                                            "$ref": "/components/schemas/candidatPostuler"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur ou formation non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé ou Formation n'existe pas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/formations/candidatureList": {
            "get": {
                "tags": [
                    "Formations"
                ],
                "summary": "Liste des candidatures avec informations",
                "description": "Récupère la liste de toutes les candidatures avec des informations détaillées sur la formation et l'utilisateur associé.",
                "operationId": "2919d7fab60de050185b70742b60d1a0",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "candidatures": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "formation": {
                                                        "$ref": "/components/schemas/Formations"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nom": {
                                                                "type": "string",
                                                                "example": "John"
                                                            },
                                                            "prenom": {
                                                                "type": "string",
                                                                "example": "Doe"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "john@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidatureList/{candidatPost}": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Modifier le statut d'une candidature",
                "description": "Modifie le statut d'une candidature de 'Refuser' à 'Accepter'.",
                "operationId": "3f5bff4ce82d13a40cac47a5ab812d52",
                "parameters": [
                    {
                        "name": "candidatPost",
                        "in": "path",
                        "description": "ID de la candidature à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut de la candidature modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Statut de la candidature modifié avec succès"
                                        },
                                        "candidature": {
                                            "$ref": "/components/schemas/candidatPostuler"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Impossible de modifier le statut de la candidature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Impossible de modifier le statut de la candidature"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Statut déjà accepté ou candidature non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Statut déjà accepté ou candidature non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidature/Accepter": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Liste des candidatures acceptées",
                "description": "Récupère la liste de toutes les candidatures avec le statut 'Accepter'.",
                "operationId": "363aa6ac913b0238848ec1694189ac4a",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures acceptées récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "listes des candidature Accepter"
                                        },
                                        "candidatures": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "/components/schemas/candidatPostuler"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidature/Refuser": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Liste des candidatures refusées",
                "description": "Récupère la liste de toutes les candidatures avec le statut 'Refuser'.",
                "operationId": "98ccc140380cbeae8765a6442b81f3c3",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures refusées récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des candidature Refuser"
                                        },
                                        "candidatures": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "/components/schemas/candidatPostuler"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}